Refactor dashboard module into package
